연재 페이지 css 정리

+++ khbooks/src/main/java/dto/BookDTO.java
@@ -9,7 +9,7 @@ public class BookDTO {
 	private String bname;
 	private int gno;
 	private Date bupdate;
-	private int bgrade;
+	private float bgrade;
 	private int bview;
 	private String bthumb;
 	private String binfo;
@@ -59,10 +59,10 @@ public class BookDTO {
 	public void setBupdate(Date bupdate) {
 		this.bupdate = bupdate;
 	}
-	public int getBgrade() {
+	public float getBgrade() {
 		return bgrade;
 	}
-	public void setBgrade(int bgrade) {
+	public void setBgrade(float bgrade) {
 		this.bgrade = bgrade;
 	}
 	public int getBview() {
