+++ khbooks/src/main/java/dto/SerialDTO.java
@@ -13,7 +13,7 @@ public class SerialDTO {
 	private int ncount;
 	private int auno;
 	private String stitle;
-	private int sgrade;
+	private float sgrade;
 	private MultipartFile filename;
 	
 	public MultipartFile getFilename() {
@@ -70,10 +70,10 @@ public class SerialDTO {
 	public void setStitle(String stitle) {
 		this.stitle = stitle;
 	}
-	public int getSgrade() {
+	public float getSgrade() {
 		return sgrade;
 	}
-	public void setSgrade(int sgrade) {
+	public void setSgrade(float sgrade) {
 		this.sgrade = sgrade;
 	}
 }
